Move `Widgets.ChromeLogic` to its own file
