+++ OpenRA.Game/Widgets/ChromeLogic.cs
@@ -1,0 +1,24 @@
+#region Copyright & License Information
+/*
+ * Copyright (c) The OpenRA Developers and Contributors
+ * This file is part of OpenRA, which is free software. It is made
+ * available to you under the terms of the GNU General Public License
+ * as published by the Free Software Foundation, either version 3 of
+ * the License, or (at your option) any later version. For more
+ * information, see COPYING.
+ */
+#endregion
+
+using System;
+
+namespace OpenRA.Widgets
+{
+	public class ChromeLogic : IDisposable
+	{
+		public void Dispose() { Dispose(true); GC.SuppressFinalize(this); }
+		public virtual void Tick() { }
+		public virtual void BecameHidden() { }
+		public virtual void BecameVisible() { }
+		protected virtual void Dispose(bool disposing) { }
+	}
+}
